feat: Fix multiple issues and improve widget functionality
This change addresses several issues in the app and improves the overall codebase.

Fixes the following functional issues:
- Today tab is empty on first launch after granting location permissions.
- Widget does not update when app settings are changed.
- Widget text does not resize properly.
- Widget style does not match the app.

Also includes the following improvements:
- Consolidates broadcast receivers to a single receiver.
- Uses constants for broadcast actions.

Additionally, this change fixes several compilation errors and lint warnings:
- Resolves type mismatches in MainViewModel, PrayerTimesRepository, and TodayScreen.
- Adds missing constants required for building.
- Adds the POST_NOTIFICATIONS permission to the manifest and handles it in the code to fix lint warnings.
- Makes implicit intents explicit to resolve lint warnings about unsafe intent launches.

+++ app/src/main/java/islam/adhanalarm/ui/TodayScreen.kt
@@ -89,7 +89,7 @@ fun DefaultPreview() {
         Array(7) { java.util.GregorianCalendar() },
         BooleanArray(7) { false },
         fi.joensuu.joyds1.calendar.IslamicCalendar(),
-        CONSTANT.DHUHR
+        CONSTANT.DHUHR.toShort()
 
     )
     PrayerTimeList(scheduleData = dummyScheduleData)
